chore: add missing files
