+++ src/contexts/index.ts
@@ -6,8 +6,13 @@
 export { default as createMagentoExtensionCtx } from "./extension";
 export { default as createGlobalCtx } from "./global";
 export { default as createProductCtx } from "./product";
-export * from "./recommendations";
-export * from "./search";
+export { default as createRecommendationUnitCtx } from "./recommendationUnit";
+export { default as createRecommendedItemCtx } from "./recommendedItem";
+export { default as createSearchInputCtx } from "./searchInput";
+export { default as createSearchResultCategoryCtx } from "./searchResultCategory";
+export { default as createSearchResultProductCtx } from "./searchResultProduct";
+export { default as createSearchResultsCtx } from "./searchResults";
+export { default as createSearchResultSuggestionCtx } from "./searchResultSuggestion";
 export { default as createShopperCtx } from "./shopper";
 export { default as createShoppingCartCtx } from "./shoppingCart";
 export { default as createStorefrontInstanceCtx } from "./storefront";
