--- src/contexts/recommendations/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-export { default as createRecommendationUnitCtx } from "./recommendationUnit";
-export { default as createRecommendedItemCtx } from "./recommendedItem";
