--- src/contexts/search/index.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-export { default as createSearchInputCtx } from "./searchInput";
-export { default as createSearchResultCategoryCtx } from "./searchResultCategory";
-export { default as createSearchResultProductCtx } from "./searchResultProduct";
-export { default as createSearchResultsCtx } from "./searchResults";
-export { default as createSearchResultSuggestionCtx } from "./searchResultSuggestion";
