+++ tests/handlers/recommendations/itemClick.test.ts
