+++ tests/handlers/search/categoryClick.test.ts
