+++ tests/handlers/search/productClick.test.ts
