+++ tests/handlers/search/requestSent.test.ts
