+++ tests/handlers/search/suggestionClick.test.ts
